Moved type into components
